digital rain added

+++ testing/include/ball.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stdint.h>
+#include <pthread.h>
 
 #define MAX_BALLS 10
 
@@ -9,6 +10,6 @@ typedef struct Balls {
     int       _stopSigs[MAX_BALLS];
 } Balls;
 
-int pushBall(Balls *balls, uint16_t *map, float r, float g, float b);
+int pushBall(Balls *balls, uint16_t *map, double r, double g, double b);
 
 int popBall(Balls *balls);
